pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "blurify"
version = "0.1.0"
description = "Local, privacy-first PII redaction pipeline for scanned images and PDFs"
authors = [{name = "Blurify Team"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "Pillow>=8.3.0",
    "scipy>=1.7.0",
    "easyocr>=1.6.0",
    "pytesseract>=0.3.8",
    "spacy>=3.4.0",
    "PyMuPDF>=1.20.0",
    "pdf2image>=1.17.0,<2.0.0",
    "streamlit>=1.15.0",
    "streamlit-option-menu>=0.3.0",
    "pytest>=6.0.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
presidio = [
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
]
yolo = [
    "ultralytics>=8.0.0",
]
gpu = [
    "torch>=1.13.0",
    "torchvision>=0.14.0",
]

[project.scripts]
blurify = "blurify.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["blurify*"]